added a util script; empty output tmp dir if it needs to be used by subsequent experiments
